+++ src/Ninject.Extensions.Interception.Test/Fakes/Derived.cs
@@ -9,4 +9,10 @@ namespace Ninject.Extensions.Interception.Fakes
         {
         }
     }
+
+    public class Derived2 : Base, IDerived
+    {
+        public void DoDerived()
+        { }
+    }
 }