+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- Unexpected interceptor is chosen when attribute is applied to any of the derived classes. [#26](https://github.com/ninject/Ninject.Extensions.Interception/issues/26)
+
 ## [3.3.2] - 2017-10-22
 
 ### Added